🚚 rename component files and update readme
